+++ src/main/java/com/vonage/client/voice/TransferCallPayload.java
@@ -25,16 +25,33 @@ import java.util.Objects;
 class TransferCallPayload extends ModifyCallPayload {
     private final TransferDestination destination;
 
+    /**
+     * Create a TransferCallPayload with an NCCO URL destination.
+     *
+     * @param nccoUrl URL of the NCCO to transfer the call to.
+     * @param uuid    ID of the call to transfer.
+     */
     public TransferCallPayload(String nccoUrl, String uuid) {
         super(ModifyCallAction.TRANSFER, uuid);
         destination = new TransferDestination(nccoUrl);
     }
 
+    /**
+     * Create a TransferCallPayload with an NCCO destination.
+     *
+     * @param ncco NCCO to transfer the call to.
+     * @param uuid ID of the call to transfer.
+     */
     public TransferCallPayload(Ncco ncco, String uuid) {
         super(ModifyCallAction.TRANSFER, uuid);
         destination = new TransferDestination(Objects.requireNonNull(ncco, "NCCO is required."));
     }
 
+    /**
+     * Call transfer destination.
+     *
+     * @return The destination of the transfer.
+     */
     @JsonProperty("destination")
     public TransferDestination getDestination() {
         return destination;