+++ src/main/java/com/vonage/client/voice/WebSocketEndpoint.java
@@ -17,18 +17,46 @@ package com.vonage.client.voice;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.vonage.client.JsonableBaseObject;
+import com.vonage.client.users.channels.Websocket;
+import java.net.URI;
 import java.util.Map;
+import java.util.Objects;
 
 public class WebSocketEndpoint extends JsonableBaseObject implements Endpoint {
+    // TODO: Use stronger typing
     private String uri, contentType;
     @JsonProperty("headers") private Map<String, Object> headers;
 
     protected WebSocketEndpoint() {
     }
 
+    /**
+     * Create a new WebSocket Endpoint
+     *
+     * @param uri URI to the websocket, starting with {@code ws://} or {@code wss://}.
+     * @param contentType The audio MIME type.
+     * @param headers Additional headers to be sent with the request.
+     */
     public WebSocketEndpoint(String uri, String contentType, Map<String, Object> headers) {
-        this.uri = uri;
-        this.contentType = contentType;
+        this(
+            URI.create(uri),
+            contentType != null ? Websocket.ContentType.fromString(contentType) : null,
+            headers
+        );
+    }
+
+    /**
+     * Create a new WebSocket Endpoint
+     *
+     * @param uri URI to the websocket, starting with {@code ws://} or {@code wss://}.
+     * @param contentType The audio MIME type.
+     * @param headers Additional headers to be sent with the request.
+     *
+     * @since 8.17.0
+     */
+    public WebSocketEndpoint(URI uri, Websocket.ContentType contentType, Map<String, Object> headers) {
+        this.uri = Objects.requireNonNull(uri, "URI is required.").toString();
+        this.contentType = contentType != null ? contentType.toString() : null;
         this.headers = headers;
     }
 
@@ -53,6 +81,7 @@ public class WebSocketEndpoint extends JsonableBaseObject implements Endpoint {
     }
 
     /**
+     * Content type of the audio stream; either {@code audio/l16;rate=16000} or {@code audio/l16;rate=8000}.
      *
      * @return The content type.
      */
@@ -61,6 +90,11 @@ public class WebSocketEndpoint extends JsonableBaseObject implements Endpoint {
         return contentType;
     }
 
+    /**
+     * Additional headers to be sent with the request.
+     *
+     * @return A map of headers to be sent in the payload.
+     */
     @JsonProperty("headers")
     public Map<String, ?> getHeadersMap() {
         return headers;